Rearange files + some renaming
