+++ crates/burn-train/src/learner/strategies/mod.rs
@@ -1,7 +1,6 @@
 mod base;
-mod multi_device;
-mod single_device;
+
+pub(crate) mod multi;
+pub(crate) mod single;
 
 pub use base::*;
-pub(crate) use multi_device::*;
-pub(crate) use single_device::*;
