+++ crates/burn-train/src/learner/strategies/multi/mod.rs
@@ -1,0 +1,4 @@
+pub(crate) mod epoch;
+mod method;
+
+pub use method::*;
